requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
openai==1.30.1
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.128.0
